Create and Remove Map Pins
Ability to create map pins in the create view, a sheet shows up and asks for a description and image. When pressing okay, it adds it to the map of which can then be viewed on all maps. You can also remove these pins but only on the create map page.