Location Visited Value
There is now a value in the array called "visited". its an int that stores if the user has been there or not, currently they can just select if they have been there or not. They pin goes green if they have selected visted. Should make it so it shows visited when entering the circle of the pin.