Added Quiz Aspect
When creating a point of interest on the map you have the options of adding a quiz. This has the question and four answers, you have to select which one is the correct asnwer.

When going onto the map, you can see the questions and clicking the asnwers make it show if it was correct or incorrect with a corresponding sounds.