Quiz Complete and Emergency Page
Made a variable that says if the quiz has been completed or not. If it has successfully been completed it will remove the quiz and show "quiz completed".  There is a bug however, that if you leave the view and come back, it will not show as completed.

Emergency page now takes you to relevant pages. Contact parents prompts to call a number, find parents takes you to find my, find security takes you to apple maps where ‘security’ is the search prompt and rescue shows a prompt that shows you how to call 000